Rename userAction.js to authAction.js
